extract shared message helpers
